Moving App.css to style folder
